FIX: Don't show SecurityID in the URL.
Showing the SecurityID in the URL is a bit of a security no-no.  Instead, we pass
it into the POST request as part of the ajax.  Note that this means that this request
requires JavaScript; in the scheme of things that's probably a slight benefit, given
that this action is a delete action.  We wouldn't want it accidentally followed!