BUG Fixing validation of past dates not being allowed.
If setting a past publish date, the job should just be triggered to
publish right now.

This is how WorkflowEmbargoExpiryExtension::onBeforeWrite() will work,
so it doesn't make sense to validate this.

Note this also fixes the issue where someone sets a publish date of
today, and hits "Now" for time, but validation fails because the
desired time has already passed by the time the person hits save.