FIX: Do not allow users with only Edit permissions to publish
By default the permission check on SiteTree::canPublish, which
canWorkflowPublish uses, falls through an ADMIN check to also
check the current user against the _edit_ permission. This has
an unintended side effect of allowing anyone with edit permission
to be able to publish when a workflow may be intended to prevent
this.