Only mark fields as readonly (not form)
Done for performance reasons. CMSMain and LeftAndMain's getEditForm()
perform this readonly transformation on Form->Fields(), not the Form itself.
It's ambiguous if the form is readonly when its fields and actions are readonly.
The easy way out is to simply apply this to the fields instead.
This should also ensure that any actions which *are* still available
for a record with an active approval state (for example "cancel approval request").
See https://github.com/dnadesign/silverstripe-elemental/issues/630
and https://github.com/silverstripe/silverstripe-framework/pull/8853