NEW Extend new AbstractGridFieldComponent class
This makes the `GridFieldExportAction` and `GridFieldWorkflowRestrictedEditButton` component `Injectable`, and allows any future enhancements in the new abstract class to automatically apply without requiring additional changes in this module.

The class is introduced in silverstripe/framework 4.11.0 so the dependency constraint needs to be updated.

Also use dependency injection instead of the `new` keyword when instantiating GridFields, their components, and configs.